Remove redundant numbering from decode type entries
We only need to number unknown and unused to make the first real
entry start from 1 and the next entries will have the next value.
Also reordered to try and keep the order the same as the original
IRremote lib and added a comment that it's somewhat important to
save that order to keep consistency with projects saving the type value
for later usage.
Note that there is still a small mismatch because this lib already has the
additional protocol COOLIX and the original IRremove has the additional
protocols as well.

Signed-off-by: Roi Dayan <roi.dayan@gmail.com>