Allow disabling the use of delay() calls. (#450)
Per Issue #430, the use of delay() calls is incompatible with the AsyncWebserver
library. Give users of the library a way to send IR messages without the delay()
calls which cause it to break the callbacks of AsyncWebserver. This is not
without risk as the delay() calls feed the WDT.

No delay() calls means the user can easily cause the WDT to reset the ESP8266.
Thus the default is to use delay() calls where possible, and make this a
compile-time option to try to limit people shooting themselves in the foot.