Samsung A/C fixes & improvements (#529)
* [Bug] Correct the byte used for Samsung AC Swing.
Wrong byte was used. Fixed.
Add extra data/unit tests to confirm.
Note: The new data highlights an issue that the lowest bit in state[9]
      changes without explaination. @Hallako requested to find out why.
* Support extended length Samsung A/C messages.
* Support 21-byte Samsung A/C Messages. e.g. Special on/off messages.
* Unit tests to support and verify new message types.
* Tune the test/Makefile
* Correct a few minor issues/style elements.
Note: Checksums for extended messages don't seem to work all the time.
* [BUG] Fix not sending Samsung A/C messages in IRMQTTServer.
* Add SAMSUNG_AC to switch statement.
* Add IRSamsungAc::sendExtended() to send extended messages.
This should handle situations when the extended/long(er) on/off
messages are to be sent from inside the IRSamsungAc class.
Ref: https://github.com/markszabo/IRremoteESP8266/issues/505#issuecomment-424591269
* More testing for Samsung AC.
* Added unit tests to decode and build states from recorded examples.
TODO(Hollako): Explain why sometimes the LSB of remote_state[9] is a 1.
e.g. 0xAE or 0XAF for swing move.
This seems to be the only thing I can't reconstruct to be the same from
all of the data provided thus far.

Ref #505