Set MWM footer and refactor state field
MWM devices typically wait 30ms between messages, so increase
footer to that length.
We're re-using the state field for MWM messages, but since payload
field is 4-bits with a minimum of 3 bytes, we only need 18 bytes.
Reworked definition of state field in union to use only what we really
need for what's defined.