Fix decoding of Samsung A/C Extended messages. (#610)
* Fix decoding of Samsung A/C Extended messages.
Turns out we had some incorrect assumptions on Samsung extended
messages. Removed some of the signature checking and make sure we pass
on the true length of the state in IRrecvDumpV2.
This should allow proper decoding of Samsung A/C Extended messages.
* Add countBits() to IRutils.
* Attempt to improve SamsungAc checksum calcs.
Checksum calculation and extended message creation are still broken in
the Samsung class. Needs more work still.

Ref #604