Create example code for a Dumb IR repeater. (#737)
The program endlessly listens for and repeats back any IR messages it gets.

It is dumb in that it doesn't try to understand the messages. It just
repeats back what ever it heard. It only replays back at a fixed
frequency.
e.g. a 40kHz signal will be re-transmitted at 38kHz.

It should work for most messages however.

This is NOT meant to be useful code, it is just example code.