Add sanity tests for unexpected conditions in IRrecv. Add some sanity checks (`assert()`) for conditions where we may have typoed a value. They only run during Unit Testing, so no performance hit.
This uncovered a situation in `decodeSony` where we could underflow an
integer (and thus make a poor match determination. i.e. No security risk)
for a hypotheically long version of an out-of-spec Sony IR message.
e.g. Far too many bits to be a Sony message.

Improved some of the Sony Unit Tests while here.