Adds Yamaha extension
Based on NEC support, Yamaha has certain functions on its recievers that utilise broken parity in the command being sent. This extension looks for that and decodes it as its own packet, rather than as an invalid command on an NEC packet. There are also some alternate encoding functions for other ways of passing data in.