Mitsubishi ACs: Improve handling swing/vane settings. (#1401)
* MitsubishiAC class
  - Tweak IRac/stdAc handing of swingv/vane:
    * Native Auto as stdAc's Off. 
    * Native Swing as stdAc's Auto.

Native "auto" is the A/C works out where the vanes point.
Native "swing" is the A/C just occilates the vanes up and down constant. Thus "swing" is closest to stdAc's "auto". Mapping the unused stdAc's "off" to the native "auto" so it is at least reachable by the user via `IRac` etc. 

* Various Mitsubish A/Cs
  - Improve `toString()` output for swingv & swingh positions.
  - Reduce code duplication.

Fixes #1399