Add Mitsubishi AC "fan only" mode (#1527)
This PR adds "fan only" mode for Mitsubishi AC units. I noticed that it was missing previously and always defaulted to "Auto" instead.

The IR code for this was captured from the remote that came with my indoor unit (Mitsubishi MSZ-SF25VE) and the resulting change to the library has been real-world tested with that device as well. I can't say if it will apply to all Mitsubishi AC units, as I only have this type available for testing.

The commits mostly just add the kMitsubishiAcFan constant and raw bitfield data.
Adds functionality to set the mode of Mitsubishi air conditioners to "fan only". Tested with MSZ-SF25VE indoor units.