Sync the on state to power from mode for HA (#1946)
Co-authored-by: Sillyfrog <sillyfrog@users.noreply.github.com>

I'm setting up the IRMQTTserver to link with HA (first time I've done it with an AC), and running with 2 outputs. I've found that the On state was doing weird things and this is the solution that worked for me (forcing the sync of the power state to On as well).

Please let me know if I missed something - it's been a while since I've had a chance to work on such things :)

I'm using the FUJITSU_AC protocol (not sure it matters, but this one has a completely different "Off" format vs On).